# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Gatherly API
  version: 0.1.0
  description: |
    The implemented API through the Phase 6 SSE increment. PostgreSQL is the
    durable source of truth; Redis accelerates disposable cache, throttling,
    and cross-instance realtime wake-ups.

    Identity is deliberately minimal: sign up or sign in with a username and
    password, then send the returned JWT in `Authorization: Bearer <token>` for
    protected endpoints. There is no development-only identity header.
servers:
  - url: http://localhost:3000
...
paths:
...
  /api/communities:
    get:
      tags: [Communities]
      summary: List active public communities
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Paginated communities.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CommunityPage' }
        '400': { $ref: '#/components/responses/Error' }
    post:
      tags: [Communities]
      summary: Create a community and an OWNER membership for the caller
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateCommunity' }
      responses:
        '201':
          description: Community created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CommunityResponse' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
        '409': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token returned by `/auth/sign-up` or `/auth/sign-in`.
  parameters:
...
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string, format: uuid }
            details: {}
      example:
        error:
          code: VALIDATION_ERROR
          message: Request validation failed
          requestId: 080f081c-d9a5-4a8d-8dd8-1d74f0137e55
          details:
            issues:
              - path: body.capacity
                message: 'Too small: expected number to be >0'
...
    Community:
      type: object
      required: [id, name, slug, description, city, country, createdByUserId, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        description: { type: string }
        city: { type: string, nullable: true }
        country: { type: string, nullable: true }
        createdByUserId: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CommunityResponse:
      type: object
      required: [data]
      properties: { data: { $ref: '#/components/schemas/Community' } }
    CommunityPage:
      type: object
      required: [data, pagination]
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/Community' } }
        pagination: { $ref: '#/components/schemas/Pagination' }
    CreateCommunity:
      type: object
      additionalProperties: false
      required: [name, slug]
      properties:
        name: { type: string, minLength: 3, maxLength: 100 }
        slug: { type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$' }
        description: { type: string, maxLength: 2000, default: '' }
        city: { type: string, minLength: 1, maxLength: 100 }
        country: { type: string, minLength: 2, maxLength: 100 }
...
    Pagination:
      type: object
      required: [page, limit, total]
      properties:
        page: { type: integer, minimum: 1 }
        limit: { type: integer, minimum: 1 }
        total: { type: integer, minimum: 0 }
...
  responses:
    Error:
      description: Safe machine-readable error response.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    AuthenticationRequired:
      description: A bearer token is missing, malformed, expired, invalid, or belongs to a non-active/deleted user (`AUTHENTICATION_REQUIRED`).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...